+++ src/app/all-posts/page.tsx
@@ -4,35 +4,20 @@ import { useState } from 'react'
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Input } from "@/components/ui/input"
 import { Button } from "@/components/ui/button"
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table"
 import { Calendar as CalendarIcon, Search } from 'lucide-react'
 import { format } from "date-fns"
 import { Calendar } from "@/components/ui/calendar"
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
 import { cn } from "@/lib/utils"
 import Header from '@/features/components/Header'
-
-// Dummy data for demonstration
-const allPosts = [
-    { id: 1, date: '2023-06-01', poster: '田中太郎', brand: 'ブランドA', itemName: '商品1', status: '査定中' },
-    { id: 2, date: '2023-06-02', poster: '鈴木花子', brand: 'ブランドB', itemName: '商品2', status: '完了' },
-    { id: 3, date: '2023-06-03', poster: '佐藤次郎', brand: 'ブランドC', itemName: '商品3', status: '査定中' },
-    { id: 4, date: '2023-06-04', poster: '山田優子', brand: 'ブランドD', itemName: '商品4', status: '完了' },
-    { id: 5, date: '2023-06-05', poster: '田中太郎', brand: 'ブランドE', itemName: '商品5', status: '査定中' },
-]
+import ItemsList from '@/features/ItemsList'
 
 export default function AllPosts() {
     const [searchDate, setSearchDate] = useState<Date | undefined>()
     const [searchPoster, setSearchPoster] = useState("")
-    const [filteredPosts, setFilteredPosts] = useState(allPosts)
 
     const handleSearch = () => {
-        const filtered = allPosts.filter(post => {
-            const dateMatch = !searchDate || post.date === format(searchDate, "yyyy-MM-dd")
-            const posterMatch = !searchPoster || post.poster.includes(searchPoster)
-            return dateMatch && posterMatch
-        })
-        setFilteredPosts(filtered)
+        // 処理を記入
     }
 
     return (
@@ -90,38 +75,7 @@ export default function AllPosts() {
                             <CardTitle className="text-2xl font-bold text-indigo-900">投稿一覧</CardTitle>
                         </CardHeader>
                         <CardContent>
-                            <div className="overflow-x-auto">
-                            <Table>
-                                <TableHeader>
-                                <TableRow>
-                                    <TableHead>ID</TableHead>
-                                    <TableHead>投稿日</TableHead>
-                                    <TableHead>投稿者</TableHead>
-                                    <TableHead>ブランド</TableHead>
-                                    <TableHead>商品名</TableHead>
-                                    <TableHead>ステータス</TableHead>
-                                </TableRow>
-                                </TableHeader>
-                                <TableBody>
-                                {filteredPosts.map((post) => (
-                                    <TableRow key={post.id}>
-                                    <TableCell>{post.id}</TableCell>
-                                    <TableCell>{post.date}</TableCell>
-                                    <TableCell>{post.poster}</TableCell>
-                                    <TableCell>{post.brand}</TableCell>
-                                    <TableCell>{post.itemName}</TableCell>
-                                    <TableCell>
-                                        <span className={`px-2 py-1 rounded-full text-xs font-semibold ${
-                                        post.status === '完了' ? 'bg-green-100 text-green-800' : 'bg-yellow-100 text-yellow-800'
-                                        }`}>
-                                        {post.status}
-                                        </span>
-                                    </TableCell>
-                                    </TableRow>
-                                ))}
-                                </TableBody>
-                            </Table>
-                            </div>
+                            <ItemsList />
                         </CardContent>
                     </Card>
                 </div>
